결제 취소 기능 구현

+++ src/main/java/com/example/demo/payment/dto/RequestPayDto.java
@@ -12,5 +12,5 @@ public class RequestPayDto {
     private String orderUid;
     private String itemName;
     private String buyerName;
-    private int paymentPrice;
+    private Long paymentPrice;
 }
